Changed selection of thermostat during run, now done with an enumerated list so we don't have to do string comparisson at each step. Also set type of the array holding the charge information to REAL8 (and corresponding changes in init_nodes) as a basis to have all real types being controlled by a single descriptor (wp8) in the future.

+++ src/qatom.f90
@@ -47,7 +47,7 @@ implicit none
 	integer(AI), allocatable							::	iqexpnb(:), jqexpnb(:)
 
 	integer									::	nqlib
-	real(4), allocatable		::	qcrg(:,:)
+	real(8), allocatable		::	qcrg(:,:)
 	real(8), allocatable		::	qmass(:)
 	real(8), allocatable		::	qavdw(:,:), qbvdw(:,:)
